requirements.txt
# Core data science dependencies
pandas>=1.5.0,<2.3.0
numpy>=1.21.0,<2.0.0
scipy>=1.9.0,<1.14.0
scikit-learn>=1.1.0,<1.5.0
matplotlib>=3.5.0,<3.9.0
seaborn>=0.11.0,<0.14.0
plotly>=5.10.0,<5.20.0

# Dashboard and visualization
dash>=2.6.0,<2.15.0
dash-bootstrap-components>=1.2.0,<1.6.0
streamlit>=1.28.2,<1.30.0

# Network analysis
networkx>=2.8.0,<3.3.0
pyvis>=0.2.1,<0.3.3

# Biology and survival analysis
lifelines>=0.27.0,<0.28.0
biopython>=1.79,<1.84

# Machine learning and dimensionality reduction
statsmodels>=0.13.0,<0.15.0
umap-learn>=0.5.0,<0.6.0
scanpy>=1.9.0,<1.10.0
anndata>=0.8.0,<0.11.0

# Utilities
requests>=2.28.0,<2.33.0
psutil>=5.9.0,<6.1.0
jupyter>=1.0.0,<1.1.0

# API and database dependencies
fastapi>=0.100.0,<0.112.0
uvicorn>=0.20.0,<0.31.0
pydantic>=2.0.0,<2.8.0
sqlalchemy>=2.0.0,<2.1.0
alembic>=1.10.0,<1.14.0
psycopg2-binary>=2.9.0,<2.10.0

# Authentication and security
passlib>=1.7.4,<1.8.0
bcrypt>=4.0.0,<4.3.0
PyJWT>=2.6.0,<2.9.0
python-multipart>=0.0.6,<0.0.8

# Background task processing
celery>=5.2.0,<5.4.0
redis>=4.3.0,<5.1.0

# Additional utilities
PyYAML>=6.0,<6.1.0
reportlab>=3.6.0,<4.1.0

# Packaging compatibility
packaging>=16.8,<24.0

# Development dependencies
pytest>=7.0.0,<8.1.0
pytest-cov>=4.0.0,<5.1.0
black>=22.0.0,<24.5.0
isort>=5.10.0,<5.14.0
flake8>=5.0.0,<7.1.0
mypy>=0.991,<1.10.0
pre-commit>=2.20.0,<3.8.0